Refactor: Reorganize project structure
