+++ src/styles/pictures/PicturesPage.module.css
