+++ src/styles/pictures/PicturesPreview.module.css
